+++ public/index.php
@@ -34,6 +34,7 @@ $router->post('/admin/users/{id}/delete', 'AdminController@deleteUser');
 $router->get('/groups/create',           'GroupController@createForm');
 $router->post('/groups/create',          'GroupController@create');
 $router->get('/groups/{id}',             'GroupController@show');
+$router->post('/groups/{id}/rename',     'GroupController@rename');
 $router->post('/groups/{id}/delete',     'GroupController@delete');
 $router->post('/groups/{id}/invite',     'GroupController@generateInvite');
 
@@ -41,6 +42,8 @@ $router->post('/groups/{id}/invite',     'GroupController@generateInvite');
 $router->get('/groups/{id}/events/create',  'EventController@createForm');
 $router->post('/groups/{id}/events/create', 'EventController@create');
 $router->get('/events/{id}',                'EventController@show');
+$router->get('/events/{id}/edit',           'EventController@editForm');
+$router->post('/events/{id}/edit',          'EventController@edit');
 $router->post('/events/{id}/delete',        'EventController@delete');
 $router->post('/events/{id}/status',        'EventController@updateStatus');
 $router->post('/events/{id}/feedback',      'EventController@saveFeedback');